requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=5.2.0=pypi_0
attrs=23.2.0=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.12.12=hecd8cb5_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
greenlet=3.0.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.1=pypi_0
jinja2=3.1.3=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
libffi=3.4.4=hecd8cb5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.4=hcec6c5f_0
numpy=1.26.3=pypi_0
openssl=3.0.12=hca72f7f_0
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.2=pypi_0
plotly=5.18.0=pypi_0
protobuf=4.25.2=pypi_0
psycopg2=2.9.9=pypi_0
psycopg2-binary=2.9.9=pypi_0
pyarrow=15.0.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
python=3.10.13=h5ee71fb_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.33.0=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
setuptools=69.0.3=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h6c40b1e_0
streamlit=1.31.0=pypi_0
tenacity=8.2.3=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.2.0=pypi_0
validators=0.22.0=pypi_0
wheel=0.41.2=py310hecd8cb5_0
xz=5.4.5=h6c40b1e_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h4dc903c_0
